+++ src/com/intellij/idea/plugin/hybris/groovy/actions/GroovyTransactionModeActionGroup.kt
@@ -19,55 +19,17 @@ package com.intellij.idea.plugin.hybris.groovy.actions
 
 import com.intellij.idea.plugin.hybris.common.utils.HybrisI18NBundleUtils.message
 import com.intellij.idea.plugin.hybris.settings.components.DeveloperSettingsComponent
-import com.intellij.openapi.actionSystem.*
+import com.intellij.idea.plugin.hybris.ui.ActionButtonWithTextAndDescription
+import com.intellij.openapi.actionSystem.ActionUpdateThread
+import com.intellij.openapi.actionSystem.AnActionEvent
+import com.intellij.openapi.actionSystem.DefaultActionGroup
 import com.intellij.openapi.actionSystem.ex.ActionUtil
-import com.intellij.openapi.actionSystem.ex.CustomComponentAction
-import com.intellij.openapi.actionSystem.impl.ActionButtonWithText
-import com.intellij.openapi.ui.popup.JBPopup
-import com.intellij.openapi.ui.popup.JBPopupFactory
-import com.intellij.ui.popup.PopupFactoryImpl
-import com.intellij.util.asSafely
-import java.awt.Dimension
-import java.io.Serial
-import javax.swing.JComponent
-import javax.swing.JList
-import javax.swing.SwingConstants
-import javax.swing.event.ListSelectionEvent
 
 class GroovyTransactionModeActionGroup : DefaultActionGroup() {
 
     init {
         templatePresentation.putClientProperty(ActionUtil.SHOW_TEXT_IN_TOOLBAR, true)
-        templatePresentation.putClientProperty(ActionUtil.COMPONENT_PROVIDER, object : CustomComponentAction {
-            override fun createCustomComponent(presentation: Presentation, place: String): JComponent = object :
-                ActionButtonWithText(this@GroovyTransactionModeActionGroup, presentation, place, Dimension()) {
-                @Serial
-                private val serialVersionUID: Long = 5346829716506322630L
-
-                override fun createAndShowActionGroupPopup(actionGroup: ActionGroup, event: AnActionEvent): JBPopup = JBPopupFactory.getInstance()
-                    .createActionGroupPopup(null, this@GroovyTransactionModeActionGroup, event.dataContext, null, true)
-                    .also { listPopup ->
-                        listPopup.listStep.values.firstOrNull()
-                            ?.asSafely<PopupFactoryImpl.ActionItem>()
-                            ?.description
-                            ?.let { listPopup.setAdText(it, 2) }
-
-                        listPopup.addListSelectionListener { e: ListSelectionEvent? ->
-                            e?.source
-                                ?.asSafely<JList<Any>>()
-                                ?.selectedValue
-                                ?.asSafely<PopupFactoryImpl.ActionItem>()
-                                ?.description
-                                ?.let { description ->
-                                    listPopup.setAdText(description, SwingConstants.LEFT)
-                                }
-                        }
-
-                        listPopup.showUnderneathOf(this)
-                    }
-            }
-        }
-        )
+        templatePresentation.putClientProperty(ActionUtil.COMPONENT_PROVIDER, ActionButtonWithTextAndDescription(this))
     }
 
     override fun getActionUpdateThread() = ActionUpdateThread.BGT