+++ src/com/intellij/idea/plugin/hybris/tools/ccv2/ui/CCv2SubscriptionsComboBoxModel.kt
@@ -1,6 +1,6 @@
 /*
  * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
- * Copyright (C) 2019-2024 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ * Copyright (C) 2019-2025 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -35,6 +35,12 @@ class CCv2SubscriptionsComboBoxModel(
     override fun setSelectedItem(anObject: Any?) {
         super.setSelectedItem(anObject)
         onSelectedItem?.invoke(anObject)
+    }
+
+    fun refresh() {
+        removeAllElements()
+        val subscriptions = ApplicationSettingsComponent.getInstance().state.ccv2Subscriptions
+        addAll(subscriptions.sortedBy { it.toString() })
     }
 
     companion object {